Fixed Multiple definitions error when compiling
* Added Makefile, util.cpp
* Modified util.hpp
